requirements.txt
pyaudiowpatch==0.2.12.7
groq==0.4.1
python-dotenv==1.0.0
numpy
noisereduce